+++ src/components/Common/index.ts
@@ -1,6 +1,5 @@
 export { default as AlertText } from './AlertText';
 export { default as Avatar } from './Avatar';
-export * from './Accordion';
 export { default as Bar } from './Bar';
 export { default as Button } from './Button';
 export { default as Checkbox } from './Checkbox';
@@ -13,21 +12,17 @@ export * from './Icon';
 export * from './IconButton';
 export * from './ImageWithFallback';
 export * from './IntersectionArea';
-export { default as Label } from './Label';
 export { default as Logo } from './Logo';
 export { default as Modal } from './Modal';
 export { default as NumberInput } from './NumberInput';
 export * from './PageHead';
 export * from './PageHeadingText';
 export { default as ProgressBar } from './ProgressBar';
-export { default as SelectBox } from './SelectBox';
 export * from './Separator';
 export { default as SkillIcon } from './SkillIcon';
 
 export { default as SsafyIcon } from './SsafyIcon';
 export * from './SsafyIcon';
-
-export { default as SSRSafeSuspense } from './SSRSafeSuspense';
 
 export { default as Tabs } from './Tabs';
 export { default as TextInput } from './TextInput';
