+++ src/components/Forms/StudentCertificationForm/Fields/Track/index.tsx
@@ -1,7 +1,7 @@
 import { css } from '@emotion/react';
 import { useId, useMemo } from 'react';
 
-import { SelectBox } from '~/components/Common';
+import { SelectBox } from '~/components/Common/SelectBox';
 import { useStudentCertificationFormContext } from '~/components/Forms/StudentCertificationForm/utils';
 import { SsafyTrack } from '~/services/member/utils';
 import { flex, fontCss } from '~/styles/utils';
