+++ src/components/Lunch/LunchCampusSelectBox/index.tsx
@@ -1,7 +1,7 @@
 import { css } from '@emotion/react';
 import React from 'react';
 
-import { SelectBox } from '~/components/Common';
+import { SelectBox } from '~/components/Common/SelectBox';
 
 interface LunchCampusSelectBoxProps {
   className?: string;
